+++ Sources/CodexBar/StatusItemController+Menu.swift
@@ -1348,6 +1348,13 @@ extension StatusItemController {
                 submenu: costSubmenu,
                 width: width))
         }
+        if !hasCredits, webItems.hasCreditsHistory, self.settings.showOptionalCreditsAndExtraUsage {
+            addSectionSeparator()
+            _ = self.addCreditsHistorySubmenu(to: menu)
+        }
+        if !hasCredits, webItems.canShowBuyCredits {
+            menu.addItem(self.makeBuyCreditsItem())
+        }
     }
 
     private func switcherIcon(for provider: UsageProvider) -> NSImage {